Refactor analytics package
