+++ src/__init__.py
@@ -1,0 +1,1 @@
+"""Core package for ContextKeeper v3 source modules."""
